user can only delete his comment and delete commentLikesRaw before delete the comment

+++ core-api/src/main/java/com/hypertube/core_api/service/CommentService.java
@@ -6,10 +6,18 @@ import com.hypertube.core_api.mapper.CommentLikesMapper;
 import com.hypertube.core_api.mapper.CommentMapper;
 import com.hypertube.core_api.model.CommentEntity;
 import com.hypertube.core_api.model.CommentLikesEntity;
+import com.hypertube.core_api.model.UserEntity;
 import com.hypertube.core_api.repository.CommentLikesRepository;
 import com.hypertube.core_api.repository.CommentRepository;
+import com.hypertube.core_api.repository.UserRepository;
+import com.hypertube.core_api.security.JwtTokenUtil;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.List;
 
 @Service
 public class CommentService {
@@ -18,16 +26,20 @@ public class CommentService {
     private final CommentMapper commentMapper;
     private final CommentLikesRepository commentLikesRepository;
     private final CommentLikesMapper commentLikesMapper;
+    private final JwtTokenUtil jwtTokenUtil;
+    private final UserRepository userRepository;
 
     public CommentService(CommentRepository commentRepository,
                           CommentMapper commentMapper,
                           CommentLikesRepository commentLikesRepository,
-                          CommentLikesMapper commentLikesMapper) {
+                          CommentLikesMapper commentLikesMapper, JwtTokenUtil jwtTokenUtil, UserRepository userRepository) {
 
         this.commentRepository = commentRepository;
         this.commentMapper = commentMapper;
         this.commentLikesRepository = commentLikesRepository;
         this.commentLikesMapper = commentLikesMapper;
+        this.jwtTokenUtil = jwtTokenUtil;
+        this.userRepository = userRepository;
     }
 
     public CommentDTO postComment(CommentDTO commentDTO) {
@@ -45,13 +57,20 @@ public class CommentService {
         return commentMapper.map(commentRepository.save(existingComment));
     }
 
-    public void deleteComment(Integer commentId) {
+    public void deleteComment(Integer commentId, String token) {
         if (!commentRepository.existsById(commentId)) {
             throw new EntityNotFoundException("Comment not found");
         }
-        CommentEntity comment = commentRepository.findById(commentId).orElseThrow(() -> new EntityNotFoundException("Comment not found"));
-        commentLikesRepository.deleteAllByCommentId(comment);
+        String username = jwtTokenUtil.extractUsername(token.replace("Bearer ", ""));
+        UserEntity userEntity = userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
+        CommentEntity comment = commentRepository.findById(commentId).orElseThrow();
 
+        if (!userEntity.getId().equals(comment.getUserId().getId())) {
+            throw new AccessDeniedException("You are not allowed to delete this comment");
+        }
+        List<CommentLikesEntity> commentLikes = commentLikesRepository.getCommentLikesEntitiesByCommentId(comment);
+        commentLikesRepository.deleteAll(commentLikes);
         commentRepository.deleteById(commentId);
     }
